pylint cleanup
