+++ tests/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 
 from ciprs_reader.parser.state import ParserState
+from ciprs_reader.parser.offense import Offenses, Offense
 
 
 @pytest.fixture
@@ -11,3 +12,13 @@ def report():
 @pytest.fixture
 def state():
     return ParserState()
+
+
+@pytest.fixture
+def offense():
+    return Offense()
+
+
+@pytest.fixture
+def offenses():
+    return Offenses()
